implement queue
Implement a queue mechanism to make sure we only send one socket request
/ response at a time.

The SocketClient is also given a better API. Before the response was
missed. Now the client emits an observable which returns with the
response.

With this new behaviour we can emit a request to poll the relay after a
relay action is executed.

We can also periodically ping a relay without fear of colliding with
other socket requests and we will be able to receive the response of the poll.